requirements.txt
claude-agent-sdk>=0.1.17
python-dotenv>=1.0.0
